feat: split all patches into separate files like a good project
