+++ src/Enhancer/Enums/ProtectionType.cs
@@ -1,0 +1,8 @@
+namespace Lethal_Company_Enhancer.Enums;
+
+public enum ProtectionType
+{
+    SAVE_NONE,
+    SAVE_ALL,
+    SAVE_COINFLIP
+}
